Fix compatibility with chprintf that doesn't support "%z"

+++ tests/common/scenarios/test_basic_connection.c
@@ -482,15 +482,15 @@ static int test_data_exchange(void) {
   int i;
   ssize_t sent;
 
-  test_printf("Primary sending %zu bytes...\n", msg_len*10);
+  test_printf("Primary sending %u bytes...\n", (uint32_t)(msg_len*10));
   for (i = 0; i < 10; ++i) {
     sent = ioHdlcWriteTmo(&peer_at_primary, test_msg, msg_len, 500);
     if (sent != (ssize_t)msg_len) {
-      test_printf("❌ Primary write returned %zd (expected %zu), errno=%d\n", 
-                  sent, msg_len, iohdlc_errno);
+      test_printf("❌ Primary write returned %d (expected %u), errno=%d\n", 
+                  (int)sent, (unsigned int)msg_len, iohdlc_errno);
     }
     TEST_ASSERT_GOTO(sent == (ssize_t)msg_len, "Primary write failed");
-    test_printf("Primary sent %zd bytes\n", sent);
+    test_printf("Primary sent %d bytes\n", (int)sent);
   }
   ioHdlc_sleep_ms(500);
 
@@ -499,8 +499,8 @@ static int test_data_exchange(void) {
   ssize_t received;
   for (i = 0; i < 10; ++i) {
     received = ioHdlcReadTmo(&peer_at_secondary, recv_buf, msg_len, 500);
-    test_printf("Secondary read returned %zd bytes (expected %zu), errno=%d\n",
-                received, msg_len, iohdlc_errno);
+    test_printf("Secondary read returned %d bytes (expected %u), errno=%d\n",
+                (int32_t)received, (uint32_t)msg_len, iohdlc_errno);
     if (received > 0 && received <= (ssize_t)sizeof recv_buf) {
       /* Null-terminate for printing */
       recv_buf[received < (ssize_t)sizeof recv_buf ? (size_t)received : sizeof recv_buf-1] = '\0';
@@ -515,20 +515,20 @@ static int test_data_exchange(void) {
   }
   TEST_ASSERT_GOTO(received == (ssize_t)msg_len, "Secondary read failed");
   TEST_ASSERT_GOTO(memcmp(recv_buf, test_msg, msg_len) == 0, "Received data mismatch");
-  test_printf("Secondary received %zd bytes: \"%s\"\n", received, recv_buf);
+  test_printf("Secondary received %d bytes: \"%s\"\n", (int32_t)received, recv_buf);
   
   /* Secondary echoes message back to primary */
   sent = ioHdlcWriteTmo(&peer_at_secondary, recv_buf, received, 500);
   TEST_ASSERT_GOTO(sent == received, "Secondary echo write failed");
-  test_printf("Secondary echoed %zd bytes\n", sent);
+  test_printf("Secondary echoed %d bytes\n", (int32_t)sent);
   
   /* Primary receives echo */
   memset(echo_buf, 0, sizeof echo_buf);
   received = ioHdlcReadTmo(&peer_at_primary, echo_buf, 40 /*sizeof echo_buf - 1*/, 500);
-  test_printf("Primary received echo %zd bytes: \"%s\"\n", received, echo_buf);
+  test_printf("Primary received echo %d bytes: \"%s\"\n", (int32_t)received, echo_buf);
   TEST_ASSERT_GOTO(received == (ssize_t)msg_len, "Primary echo read failed");
   TEST_ASSERT_GOTO(memcmp(echo_buf, test_msg, msg_len) == 0, "Echo data mismatch");
-  test_printf("Primary received echo %zd bytes: \"%s\"\n", received, echo_buf);
+  test_printf("Primary received echo %d bytes: \"%s\"\n", (int32_t)received, echo_buf);
 
   ioHdlc_sleep_ms(200);
